split region

+++ utilities/dockered_pipelines/bamSimulator/BamSimulator_multiThreads.sh
@@ -214,9 +214,15 @@ else
     cat ${HUMAN_REFERENCE}.fai | awk -F "\t" '{print $1 "\t0\t" $2}' | awk -F "\t" '$1 ~ /^(chr)?[0-9XY]+$/' > ${parent_outdir}/genome.bed
 fi
 
-docker run --rm -v /:/mnt -u $UID -i lethalfang/somaticseq:${VERSION} \
-/opt/somaticseq/utilities/split_Bed_into_equal_regions.py \
--infile /mnt/${parent_outdir}/genome.bed -num $threads -outfiles /mnt/${parent_outdir}/bed
+
+if [[ `which python3` ]]
+then
+     $MYDIR/../../split_Bed_into_equal_regions.py -infile ${outdir}/genome.bed -num $threads -outfiles ${outdir}/bed
+else
+    docker run --rm -v /:/mnt -u $UID -i lethalfang/somaticseq:${VERSION} \
+    /opt/somaticseq/utilities/split_Bed_into_equal_regions.py \
+    -infile /mnt/${parent_outdir}/genome.bed -num $threads -outfiles /mnt/${parent_outdir}/bed
+fi
 
 
 ith_thread=1
